+++ pom.xml
@@ -10,10 +10,12 @@
     <version>1.1-SNAPSHOT</version>
 
     <properties>
+        <servlet.api>3.0.1</servlet.api>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <spring.version>RELEASE</spring.version>
         <jackson.version>RELEASE</jackson.version>
         <spring.security.version>RELEASE</spring.security.version>
+        <hibernate.version>RELEASE</hibernate.version>
     </properties>
 
     <dependencies>
@@ -21,7 +23,7 @@
         <dependency>
             <groupId>javax.servlet</groupId>
             <artifactId>javax.servlet-api</artifactId>
-            <version>3.0.1</version>
+            <version>${servlet.api}</version>
             <scope>provided</scope>
         </dependency>
 
@@ -74,6 +76,11 @@
             <groupId>org.springframework.security</groupId>
             <artifactId>spring-security-config</artifactId>
             <version>${spring.security.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.springframework.data</groupId>
+            <artifactId>spring-data-jpa</artifactId>
+            <version>RELEASE</version>
         </dependency>
 
         <dependency>
@@ -131,8 +138,14 @@
         <dependency>
             <groupId>org.hibernate</groupId>
             <artifactId>hibernate-core</artifactId>
-            <version>RELEASE</version>
+            <version>${hibernate.version}</version>
         </dependency>
+        <!--<dependency>-->
+            <!--<groupId>org.hibernate</groupId>-->
+            <!--<artifactId>hibernate-entitymanager</artifactId>-->
+            <!--<version>${hibernate.version}</version>-->
+            <!--<scope>runtime</scope>-->
+        <!--</dependency>-->
 
         <dependency>
             <groupId>org.postgresql</groupId>
