+++ src/main/java/eu/bausov/projects/srvpumpselection/dao/SpeedCorrectionCoefficientDAO.java
@@ -1,0 +1,10 @@
+package eu.bausov.projects.srvpumpselection.dao;
+
+import eu.bausov.projects.srvpumpselection.bo.SpeedCorrectionCoefficient;
+import org.springframework.data.repository.CrudRepository;
+
+/**
+ * Created by GreenNun on 18.03.17.
+ */
+public interface SpeedCorrectionCoefficientDAO extends CrudRepository<SpeedCorrectionCoefficient, Long> {
+}
