+++ src/main/resources/dataBaseContext.xml
@@ -1,7 +1,15 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <beans xmlns="http://www.springframework.org/schema/beans"
        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-       xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd">
+       xmlns:context="http://www.springframework.org/schema/context"
+       xmlns:jpa="http://www.springframework.org/schema/data/jpa"
+       xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-3.2.xsd
+           http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context-3.2.xsd
+           http://www.springframework.org/schema/data/jpa http://www.springframework.org/schema/data/jpa/spring-jpa-1.0.xsd">
+
+    <context:annotation-config/>
+    <context:component-scan base-package="eu.bausov.projects.srvpumpselection"/>
+    <context:property-placeholder location="classpath:service.properties"/>
 
     <bean id="serviceDataSource" class="org.apache.commons.dbcp.BasicDataSource" destroy-method="close">
         <property name="username" value="${core.dataSourceUsername}"/>
@@ -28,14 +36,34 @@
         <property name="accessToUnderlyingConnectionAllowed" value="true"/>
     </bean>
 
-    <bean id="sessionFactory" class="org.springframework.orm.hibernate4.LocalSessionFactoryBean" destroy-method="destroy">
-        <property name="dataSource" ref="serviceDataSource"/>
-        <property name="configLocation" value="classpath:/hibernate.cfg.xml"/>
-        <property name="annotatedPackages" value="eu.bausov.projects.srvpumpselection.bo.equipment"/>
+    <jpa:repositories base-package="eu.bausov.projects.srvpumpselection.dao"/>
+
+    <bean id="transactionManager" class="org.springframework.orm.jpa.JpaTransactionManager">
+        <property name="entityManagerFactory" ref="entityManagerFactory"/>
+        <property name="jpaDialect">
+            <bean class="org.springframework.orm.jpa.vendor.HibernateJpaDialect"/>
+        </property>
     </bean>
 
-    <bean id="transactionManager" class="org.springframework.orm.hibernate4.HibernateTransactionManager">
-        <property name="sessionFactory" ref="sessionFactory"/>
+    <bean id="entityManagerFactory" class="org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean">
+        <property name="dataSource" ref="serviceDataSource"/>
+        <property name="packagesToScan" value="eu.bausov.projects.srvpumpselection.bo"/>
+        <property name="jpaVendorAdapter">
+            <bean class="org.springframework.orm.jpa.vendor.HibernateJpaVendorAdapter">
+                <property name="generateDdl" value="true"/>
+                <property name="showSql" value="false"/>
+                <property name="databasePlatform" value="org.hibernate.dialect.PostgreSQL9Dialect"/>
+                <property name="database" value="POSTGRESQL"/>
+            </bean>
+        </property>
+        <!--<property name="jpaProperties">-->
+            <!--<value>-->
+                <!--hibernate.cache.use_second_level_cache = true-->
+                <!--hibernate.cache.region.factory_class = org.hibernate.cache.ehcache.EhCacheRegionFactory-->
+                <!--hibernate.cache.use_query_cache = true-->
+                <!--hibernate.generate_statistics = true-->
+            <!--</value>-->
+        <!--</property>-->
     </bean>
 
 </beans>
