--- src/main/resources/hibernate.cfg.xml
@@ -1,28 +1,0 @@
-<?xml version='1.0' encoding='utf-8'?>
-<!DOCTYPE hibernate-configuration PUBLIC
-        "-//Hibernate/Hibernate Configuration DTD//EN"
-        "http://www.hibernate.org/dtd/hibernate-configuration-3.0.dtd">
-
-<hibernate-configuration>
-    <session-factory>
-
-        <property name="dialect">org.hibernate.dialect.PostgreSQL9Dialect</property>
-        <property name="id.new_generator_mappings">true</property>
-        <property name="use_sql_comments">false</property>
-        <property name="format_sql">true</property>
-        <property name="show_sql">false</property>
-
-        <mapping class="eu.bausov.projects.srvpumpselection.bo.Constant"/>
-        <mapping class="eu.bausov.projects.srvpumpselection.bo.Producer"/>
-        <mapping class="eu.bausov.projects.srvpumpselection.bo.SpeedCorrectionCoefficient"/>
-        <mapping class="eu.bausov.projects.srvpumpselection.bo.equipment.DriverAssembly"/>
-        <mapping class="eu.bausov.projects.srvpumpselection.bo.equipment.Equipment"/>
-        <mapping class="eu.bausov.projects.srvpumpselection.bo.equipment.Frame"/>
-        <mapping class="eu.bausov.projects.srvpumpselection.bo.equipment.Pump"/>
-        <mapping class="eu.bausov.projects.srvpumpselection.bo.equipment.Motor"/>
-        <mapping class="eu.bausov.projects.srvpumpselection.bo.equipment.PumpAggregate"/>
-        <mapping class="eu.bausov.projects.srvpumpselection.bo.equipment.Reducer"/>
-        <mapping class="eu.bausov.projects.srvpumpselection.bo.equipment.Seal"/>
-
-    </session-factory>
-</hibernate-configuration>
