+++ src/main/webapp/WEB-INF/mvc-servlet.xml
@@ -18,7 +18,6 @@
     <security:global-method-security pre-post-annotations="enabled"/>
 
     <import resource="classpath:webServicesContext.xml"/>
-    <!--<mvc:resources mapping="" location=""/>-->
 
     <bean id="exceptionResolver" class="eu.bausov.projects.srvpumpselection.web.rest.HandlerExceptionResolver">
         <property name="order" value="1"/>
