+++ src/main/webapp/WEB-INF/web.xml
@@ -29,21 +29,6 @@
     </listener>
 
     <filter>
-        <filter-name>HibernateOpenSessionFilter</filter-name>
-        <filter-class>org.springframework.orm.hibernate4.support.OpenSessionInViewFilter</filter-class>
-    </filter>
-
-    <filter-mapping>
-        <filter-name>HibernateOpenSessionFilter</filter-name>
-        <url-pattern>/api/*</url-pattern>
-    </filter-mapping>
-
-    <filter-mapping>
-        <filter-name>HibernateOpenSessionFilter</filter-name>
-        <url-pattern>/login/*</url-pattern>
-    </filter-mapping>
-
-    <filter>
         <filter-name>springSecurityFilterChain</filter-name>
         <filter-class>org.springframework.web.filter.DelegatingFilterProxy</filter-class>
     </filter>
